# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OrderItem {
  id: number;
  name: string;
  price: number;
  quantity: number;
}

export type OrderInputDeliveryType = typeof OrderInputDeliveryType[keyof typeof OrderInputDeliveryType];


export const OrderInputDeliveryType = {
  home: 'home',
  office: 'office',
} as const;

export interface OrderInput {
  /**
     * @minLength 12
     * @maxLength 100
     */
  deviceId: string;
  /** @minLength 2 */
  name: string;
  /** @minLength 8 */
  phone: string;
  /** @minLength 2 */
  wilaya: string;
  /** @minLength 3 */
  address: string;
  deliveryType: OrderInputDeliveryType;
  /** @minimum 0 */
  deliveryFee: number;
  /** @minimum 0 */
  total: number;
  /** @minItems 1 */
  items: OrderItem[];
}

export type Order = OrderInput & {
  id: number;
  code: string;
  createdAt: string;
};

export interface RateLimitError {
  error: string;
  retryAfterHours: number;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  stock: number;
  art: string;
  /** @nullable */
  image: string | null;
  featured: boolean;
  priceVisible: boolean;
  stockVisible: boolean;
  /** @nullable */
  promotion: string | null;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  stock: number;
  art: string;
  /** @nullable */
  image?: string | null;
  featured?: boolean;
  priceVisible?: boolean;
  stockVisible?: boolean;
  /** @nullable */
  promotion?: string | null;
}

export interface ErrorEnvelope {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface UploadResponse {
  objectPath: string;
  metadata: UploadUrlRequest;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};