# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInput } from './orderInput';

export type Order = OrderInput & {
  id: number;
  code: string;
  createdAt: Date;
};