# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputDeliveryType } from './orderInputDeliveryType';
import type { OrderItem } from './orderItem';

export interface OrderInput {
  /**
     * @minLength 12
     * @maxLength 100
     */
  deviceId: string;
  /** @minLength 2 */
  name: string;
  /** @minLength 8 */
  phone: string;
  /** @minLength 2 */
  wilaya: string;
  /** @minLength 3 */
  address: string;
  deliveryType: OrderInputDeliveryType;
  /** @minimum 0 */
  deliveryFee: number;
  /** @minimum 0 */
  total: number;
  /** @minItems 1 */
  items: OrderItem[];
}