# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  stock: number;
  art: string;
  /** @nullable */
  image: string | null;
  featured: boolean;
  priceVisible: boolean;
  stockVisible: boolean;
  /** @nullable */
  promotion: string | null;
}