# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  stock: number;
  art: string;
  /** @nullable */
  image?: string | null;
  featured?: boolean;
  priceVisible?: boolean;
  stockVisible?: boolean;
  /** @nullable */
  promotion?: string | null;
}