# lib/api-zod/src/generated/types/uploadResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UploadUrlRequest } from './uploadUrlRequest';

export interface UploadResponse {
  objectPath: string;
  metadata: UploadUrlRequest;
}